[fix] rename_field path
